requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-ckeditor==6.7.0
django-allauth==0.58.2
django-environ==0.11.2
django-taggit==4.0.0
django-cleanup==8.0.0